ai/requirements.txt
fastapi==0.128.0
uvicorn[standard]
numpy==1.26.4
torchaudio
ultralytics==8.4.2
opencv-python-headless>=4.8
albumentations==1.4.10
pillow==12.1.0
scikit-learn==1.5.2
pandas==2.2.2
requests
jupyterlab
matplotlib
python-multipart
mlflow==3.8.1
boto3